Реорганизация store. Перенос селекторов
