+++ src/components/film-catalog/film-catalog.jsx
@@ -2,8 +2,7 @@ import PropTypes from 'prop-types';
 import React from "react";
 import {connect} from 'react-redux';
 import {SHOW_MORE_FILMS_STEP} from '../../const';
-import {changeCurrentGenre, increaseShownFilmsCount, resetShownFilmsCount} from '../../store/catalog/catalog';
-import {selectCurrentGenre, selectFilteredFilms, selectGenreList, selectIsAllFilmsShown, selectShownFilmsCount} from '../../store/selectors';
+import {changeCurrentGenre, increaseShownFilmsCount, resetShownFilmsCount, selectCurrentGenre, selectFilteredFilms, selectGenreList, selectIsAllFilmsShown, selectShownFilmsCount} from '../../store/catalog/catalog';
 import {filmShape} from "../../utils/props-validation";
 import FilmsList from "../films-list/films-list";
 import GenreFilter from "../genre-filter/genre-filter";
