+++ src/components/film-screen/film-screen.jsx
@@ -2,7 +2,8 @@ import PropTypes from 'prop-types';
 import React, {useEffect} from "react";
 import {connect} from 'react-redux';
 import {fetchFilmById, fetchReviewsByFilmId} from '../../store/api-actions';
-import {selectFilm, selectIsFilmLoaded, selectIsUserLogged, selectReviews, selectSimilarFilms} from '../../store/selectors';
+import {selectFilm, selectIsFilmLoaded, selectReviews, selectSimilarFilms} from '../../store/data/data';
+import {selectIsUserLogged} from '../../store/user/user';
 import {filmShape, reviewShape} from "../../utils/props-validation";
 import FilmCard from '../film-card/film-card';
 import FilmInfo from '../film-info/film-info';
