+++ src/components/my-list-screen/my-list-screen.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from "react";
 import {connect} from 'react-redux';
-import {selectFavoriteFilms} from '../../store/selectors';
+import {selectFavoriteFilms} from '../../store/data/data';
 import {filmShape} from "../../utils/props-validation";
 import FilmsList from "../films-list/films-list";
 import Footer from '../footer/footer';
