+++ src/store/namespace.js
@@ -1,0 +1,6 @@
+export const NameSpace = {
+  DATA: `DATA`,
+  CATALOG: `CATALOG`,
+  USER: `USER`,
+  REVIEW: `REVIEW`,
+};
